Move API-exposed types into dpd-types-versions
Define the remaining API wire types in their introduction-version
modules, and make dpd-types, common, and aal compatibility re-export
layers. Keep runtime/controller types out of the API crate by converting
transceiver values at the DPD boundary.
