+++ dpd-api/Cargo.toml
@@ -4,9 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-common.workspace = true
-dpd-types.workspace = true
 dpd-types-versions.workspace = true
 dropshot.workspace = true
 dropshot-api-manager-types.workspace = true
-transceiver-controller.workspace = true
