+++ dpd-api/src/lib.rs
@@ -9,13 +9,6 @@
 use std::collections::BTreeMap;
 use std::net::{Ipv4Addr, Ipv6Addr};
 
-use common::{
-    attached_subnet::AttachedSubnetEntry,
-    nat::{Ipv4Nat, Ipv6Nat},
-    network::{InstanceTarget, MacAddr, NatTarget},
-    ports::{Ipv4Entry, Ipv6Entry, PortId, PortPrbsMode, TxEq, TxEqSwHw},
-};
-use dpd_types::oxstats;
 use dpd_types_versions::{latest, v1, v4, v7};
 use dropshot::{
     EmptyScanParams, HttpError, HttpResponseCreated, HttpResponseDeleted,
@@ -23,9 +16,6 @@ use dropshot::{
     Query, RequestContext, ResultsPage, TypedBody,
 };
 use dropshot_api_manager_types::api_versions;
-use transceiver_controller::{
-    Datapath, Monitors, PowerState, message::LedState,
-};
 
 api_versions!([
     // WHEN CHANGING THE API (part 1 of 2):
@@ -500,7 +490,7 @@ pub trait DpdApi {
     }]
     async fn port_list(
         rqctx: RequestContext<Self::Context>,
-    ) -> Result<HttpResponseOk<Vec<PortId>>, HttpError>;
+    ) -> Result<HttpResponseOk<Vec<latest::port::PortId>>, HttpError>;
 
     /// Get the set of available channels for all ports.
     ///
@@ -573,7 +563,7 @@ pub trait DpdApi {
     async fn led_set(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::port::PortIdPathParams>,
-        body: TypedBody<LedState>,
+        body: TypedBody<latest::switch_port::LedState>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /// Return the full backplane map.
@@ -588,7 +578,9 @@ pub trait DpdApi {
     async fn backplane_map(
         rqctx: RequestContext<Self::Context>,
     ) -> Result<
-        HttpResponseOk<BTreeMap<PortId, latest::port_map::BackplaneLink>>,
+        HttpResponseOk<
+            BTreeMap<latest::port::PortId, latest::port_map::BackplaneLink>,
+        >,
         HttpError,
     >;
 
@@ -610,7 +602,9 @@ pub trait DpdApi {
     async fn leds_list(
         rqctx: RequestContext<Self::Context>,
     ) -> Result<
-        HttpResponseOk<BTreeMap<PortId, latest::switch_port::Led>>,
+        HttpResponseOk<
+            BTreeMap<latest::port::PortId, latest::switch_port::Led>,
+        >,
         HttpError,
     >;
 
@@ -635,7 +629,9 @@ pub trait DpdApi {
     async fn transceivers_list(
         rqctx: RequestContext<Self::Context>,
     ) -> Result<
-        HttpResponseOk<BTreeMap<PortId, latest::transceivers::Transceiver>>,
+        HttpResponseOk<
+            BTreeMap<latest::port::PortId, latest::transceivers::Transceiver>,
+        >,
         HttpError,
     >;
 
@@ -677,7 +673,7 @@ pub trait DpdApi {
     async fn transceiver_power_set(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::port::PortIdPathParams>,
-        state: TypedBody<PowerState>,
+        state: TypedBody<latest::transceivers::PowerState>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /// Return the power state of a transceiver.
@@ -688,7 +684,7 @@ pub trait DpdApi {
     async fn transceiver_power_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::port::PortIdPathParams>,
-    ) -> Result<HttpResponseOk<PowerState>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::transceivers::PowerState>, HttpError>;
 
     /// Fetch the monitored environmental information for the provided transceiver.
     #[endpoint {
@@ -698,7 +694,7 @@ pub trait DpdApi {
     async fn transceiver_monitors_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::port::PortIdPathParams>,
-    ) -> Result<HttpResponseOk<Monitors>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::transceivers::Monitors>, HttpError>;
 
     /// Fetch the state of the datapath for the provided transceiver.
     #[endpoint {
@@ -708,7 +704,7 @@ pub trait DpdApi {
     async fn transceiver_datapath_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::port::PortIdPathParams>,
-    ) -> Result<HttpResponseOk<Datapath>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::transceivers::Datapath>, HttpError>;
 
     /// Create a link on a switch port.
     ///
@@ -923,7 +919,7 @@ pub trait DpdApi {
     async fn link_prbs_set(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-        body: TypedBody<PortPrbsMode>,
+        body: TypedBody<latest::port::PortPrbsMode>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /// Set a link's PRBS speed and mode.
@@ -938,9 +934,10 @@ pub trait DpdApi {
         path: Path<v1::link::LinkPath>,
         body: TypedBody<v1::port::PortPrbsMode>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-        let mode = PortPrbsMode::try_from(body.into_inner()).map_err(|e| {
-            HttpError::for_bad_request(None, format!("bad PRBS mode: {e}"))
-        })?;
+        let mode = latest::port::PortPrbsMode::try_from(body.into_inner())
+            .map_err(|e| {
+                HttpError::for_bad_request(None, format!("bad PRBS mode: {e}"))
+            })?;
 
         Self::link_prbs_set(rqctx, path, TypedBody::from(mode)).await
     }
@@ -958,7 +955,7 @@ pub trait DpdApi {
     async fn link_prbs_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-    ) -> Result<HttpResponseOk<PortPrbsMode>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::port::PortPrbsMode>, HttpError>;
 
     /// Return the link's PRBS speed and mode.
     ///
@@ -1028,7 +1025,7 @@ pub trait DpdApi {
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
         query: Query<PaginationParams<EmptyScanParams, latest::arp::Ipv4Token>>,
-    ) -> Result<HttpResponseOk<ResultsPage<Ipv4Entry>>, HttpError>;
+    ) -> Result<HttpResponseOk<ResultsPage<latest::port::Ipv4Entry>>, HttpError>;
 
     /// Add an IPv4 address to a link.
     #[endpoint {
@@ -1038,7 +1035,7 @@ pub trait DpdApi {
     async fn link_ipv4_create(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-        entry: TypedBody<Ipv4Entry>,
+        entry: TypedBody<latest::port::Ipv4Entry>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /// Clear all IPv4 addresses from a link.
@@ -1070,7 +1067,7 @@ pub trait DpdApi {
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
         query: Query<PaginationParams<EmptyScanParams, latest::arp::Ipv6Token>>,
-    ) -> Result<HttpResponseOk<ResultsPage<Ipv6Entry>>, HttpError>;
+    ) -> Result<HttpResponseOk<ResultsPage<latest::port::Ipv6Entry>>, HttpError>;
 
     /// Add an IPv6 address to a link.
     #[endpoint {
@@ -1080,7 +1077,7 @@ pub trait DpdApi {
     async fn link_ipv6_create(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-        entry: TypedBody<Ipv6Entry>,
+        entry: TypedBody<latest::port::Ipv6Entry>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /// Clear all IPv6 addresses from a link.
@@ -1111,7 +1108,7 @@ pub trait DpdApi {
     async fn link_mac_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-    ) -> Result<HttpResponseOk<MacAddr>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::network::MacAddr>, HttpError>;
 
     /// Set a link's MAC address.
     // TODO-correctness: A link's MAC address should be determined by the FRUID
@@ -1125,7 +1122,7 @@ pub trait DpdApi {
     async fn link_mac_set(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-        body: TypedBody<MacAddr>,
+        body: TypedBody<latest::network::MacAddr>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /// Return whether the link is configured to drop non-nat traffic
@@ -1216,7 +1213,7 @@ pub trait DpdApi {
     }]
     async fn loopback_ipv4_list(
         rqctx: RequestContext<Self::Context>,
-    ) -> Result<HttpResponseOk<Vec<Ipv4Entry>>, HttpError>;
+    ) -> Result<HttpResponseOk<Vec<latest::port::Ipv4Entry>>, HttpError>;
 
     /**
      * Add a loopback IPv4.
@@ -1227,7 +1224,7 @@ pub trait DpdApi {
     }]
     async fn loopback_ipv4_create(
         rqctx: RequestContext<Self::Context>,
-        val: TypedBody<Ipv4Entry>,
+        val: TypedBody<latest::port::Ipv4Entry>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /**
@@ -1251,7 +1248,7 @@ pub trait DpdApi {
     }]
     async fn loopback_ipv6_list(
         rqctx: RequestContext<Self::Context>,
-    ) -> Result<HttpResponseOk<Vec<Ipv6Entry>>, HttpError>;
+    ) -> Result<HttpResponseOk<Vec<latest::port::Ipv6Entry>>, HttpError>;
 
     /**
      * Add a loopback IPv6.
@@ -1262,7 +1259,7 @@ pub trait DpdApi {
     }]
     async fn loopback_ipv6_create(
         rqctx: RequestContext<Self::Context>,
-        val: TypedBody<Ipv6Entry>,
+        val: TypedBody<latest::port::Ipv6Entry>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /**
@@ -1300,7 +1297,7 @@ pub trait DpdApi {
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::nat::NatIpv6Path>,
         query: Query<PaginationParams<EmptyScanParams, latest::nat::NatToken>>,
-    ) -> Result<HttpResponseOk<ResultsPage<Ipv6Nat>>, HttpError>;
+    ) -> Result<HttpResponseOk<ResultsPage<latest::nat::Ipv6Nat>>, HttpError>;
 
     /**
      * Get the external->internal NAT mapping for the given address and starting L3
@@ -1313,7 +1310,7 @@ pub trait DpdApi {
     async fn nat_ipv6_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::nat::NatIpv6PortPath>,
-    ) -> Result<HttpResponseOk<NatTarget>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::network::NatTarget>, HttpError>;
 
     /**
      * Add an external->internal NAT mapping for the given address and L3 port
@@ -1335,7 +1332,7 @@ pub trait DpdApi {
     async fn nat_ipv6_create(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::nat::NatIpv6RangePath>,
-        target: TypedBody<NatTarget>,
+        target: TypedBody<latest::network::NatTarget>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /**
@@ -1384,7 +1381,7 @@ pub trait DpdApi {
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::nat::NatIpv4Path>,
         query: Query<PaginationParams<EmptyScanParams, latest::nat::NatToken>>,
-    ) -> Result<HttpResponseOk<ResultsPage<Ipv4Nat>>, HttpError>;
+    ) -> Result<HttpResponseOk<ResultsPage<latest::nat::Ipv4Nat>>, HttpError>;
 
     /**
      * Get the external->internal NAT mapping for the given address/port
@@ -1396,7 +1393,7 @@ pub trait DpdApi {
     async fn nat_ipv4_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::nat::NatIpv4PortPath>,
-    ) -> Result<HttpResponseOk<NatTarget>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::network::NatTarget>, HttpError>;
 
     /**
      * Add an external->internal NAT mapping for the given address/port range
@@ -1418,7 +1415,7 @@ pub trait DpdApi {
     async fn nat_ipv4_create(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::nat::NatIpv4RangePath>,
-        target: TypedBody<NatTarget>,
+        target: TypedBody<latest::network::NatTarget>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /**
@@ -1460,7 +1457,12 @@ pub trait DpdApi {
                 latest::route::AttachedSubnetToken,
             >,
         >,
-    ) -> Result<HttpResponseOk<ResultsPage<AttachedSubnetEntry>>, HttpError>;
+    ) -> Result<
+        HttpResponseOk<
+            ResultsPage<latest::attached_subnet::AttachedSubnetEntry>,
+        >,
+        HttpError,
+    >;
 
     /**
      * Get the mapping for the given external subnet.
@@ -1473,7 +1475,7 @@ pub trait DpdApi {
     async fn attached_subnet_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::route::SubnetPath>,
-    ) -> Result<HttpResponseOk<InstanceTarget>, HttpError>;
+    ) -> Result<HttpResponseOk<latest::network::InstanceTarget>, HttpError>;
 
     /**
      * Add a mapping to an internal target for an external subnet address.
@@ -1490,7 +1492,7 @@ pub trait DpdApi {
     async fn attached_subnet_create(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::route::SubnetPath>,
-        target: TypedBody<InstanceTarget>,
+        target: TypedBody<latest::network::InstanceTarget>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /**
@@ -1618,7 +1620,10 @@ pub trait DpdApi {
     }]
     async fn oximeter_collect_meta_endpoint(
         rqctx: RequestContext<Self::Context>,
-    ) -> Result<HttpResponseOk<Option<oxstats::OximeterMetadata>>, HttpError>;
+    ) -> Result<
+        HttpResponseOk<Option<latest::oxstats::OximeterMetadata>>,
+        HttpError,
+    >;
 
     /**
      * Apply port settings atomically.
@@ -2566,7 +2571,7 @@ pub trait DpdApi {
     async fn link_tx_eq_get(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-    ) -> Result<HttpResponseOk<Vec<TxEqSwHw>>, HttpError>;
+    ) -> Result<HttpResponseOk<Vec<latest::port::TxEqSwHw>>, HttpError>;
 
     /**
      * Update the per-lane tx eq settings for all lanes on this link
@@ -2578,7 +2583,7 @@ pub trait DpdApi {
     async fn link_tx_eq_set(
         rqctx: RequestContext<Self::Context>,
         path: Path<latest::link::LinkPath>,
-        args: TypedBody<TxEq>,
+        args: TypedBody<latest::port::TxEq>,
     ) -> Result<HttpResponseUpdatedNoContent, HttpError>;
 
     /**
