+++ dpd-types/Cargo.toml
@@ -4,8 +4,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-chrono.workspace = true
 dpd-types-versions.workspace = true
-omicron-common.workspace = true
-schemars.workspace = true
-serde.workspace = true
