+++ dpd-types/src/attached_subnet.rs
@@ -4,10 +4,4 @@
 //
 // Copyright 2026 Oxide Computer Company
 
-//! Version `ASIC_DETAILS` of the DPD API.
-//!
-//! Added `lotnum` and `fuse` fields to `SwitchIdentifiers`.
-//! Updated `OximeterConfig` and `OximeterMetadata` to use the new identifiers.
-
-pub mod oxstats;
-pub mod switch_identifiers;
+pub use dpd_types_versions::latest::attached_subnet::*;
