+++ dpd-types/versions/Cargo.toml
@@ -4,13 +4,12 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-aal.workspace = true
-common.workspace = true
+chrono.workspace = true
 dropshot.workspace = true
 omicron-common.workspace = true
 oxnet.workspace = true
+rand.workspace = true
 schemars.workspace = true
 serde.workspace = true
 thiserror.workspace = true
-transceiver-controller = { workspace = true, features = ["api-traits"] }
 uuid.workspace = true
