+++ dpd-types/versions/src/impls/transceivers.rs
@@ -8,8 +8,18 @@ use std::time::Instant;
 
 use crate::latest::switch_port::ManagementMode;
 use crate::latest::transceivers::{
-    ElectricalMode, QsfpDevice, TransceiverInfo,
+    ElectricalMode, PowerState, QsfpDevice, TransceiverInfo,
 };
+
+impl std::fmt::Display for PowerState {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            PowerState::Off => f.write_str("off"),
+            PowerState::Low => f.write_str("low"),
+            PowerState::High => f.write_str("high"),
+        }
+    }
+}
 
 impl Default for QsfpDevice {
     fn default() -> Self {
