+++ dpd-types/versions/src/latest.rs
@@ -15,13 +15,22 @@ pub mod arp {
     pub use crate::v1::arp::Ipv6Token;
 }
 
+pub mod attached_subnet {
+    pub use crate::v3::attached_subnet::AttachedSubnetEntry;
+}
+
 pub mod counters {
+    pub use crate::v1::counters::CounterData;
     pub use crate::v1::counters::CounterPath;
     pub use crate::v1::counters::CounterSync;
+    pub use crate::v1::counters::FecRSCounters;
     pub use crate::v1::counters::LinkFecRSCounters;
     pub use crate::v1::counters::LinkPcsCounters;
     pub use crate::v1::counters::LinkRMonCounters;
     pub use crate::v1::counters::LinkRMonCountersAll;
+    pub use crate::v1::counters::PcsCounters;
+    pub use crate::v1::counters::RMonCounters;
+    pub use crate::v1::counters::RMonCountersAll;
 }
 
 pub mod fault {
@@ -89,6 +98,8 @@ pub mod misc {
 }
 
 pub mod nat {
+    pub use crate::v1::nat::Ipv4Nat;
+    pub use crate::v1::nat::Ipv6Nat;
     pub use crate::v1::nat::NatIpv4Path;
     pub use crate::v1::nat::NatIpv4PortPath;
     pub use crate::v1::nat::NatIpv4RangePath;
@@ -98,16 +109,48 @@ pub mod nat {
     pub use crate::v1::nat::NatToken;
 }
 
+pub mod network {
+    pub use crate::v1::network::MacAddr;
+    pub use crate::v1::network::MacError;
+    pub use crate::v1::network::NatTarget;
+    pub use crate::v1::network::Vni;
+
+    pub use crate::v3::network::InstanceTarget;
+}
+
+pub mod oxstats {
+    pub use crate::v1::oxstats::SledIdentifiers;
+
+    pub use crate::v10::oxstats::OximeterConfig;
+    pub use crate::v10::oxstats::OximeterMetadata;
+}
+
 pub mod port {
     pub use crate::v1::port::FreeChannels;
+    pub use crate::v1::port::InternalPort;
+    pub use crate::v1::port::Ipv4Entry;
+    pub use crate::v1::port::Ipv6Entry;
     pub use crate::v1::port::LinkSettings;
+    pub use crate::v1::port::PORT_COUNT_INTERNAL;
+    pub use crate::v1::port::PORT_COUNT_QSFP;
+    pub use crate::v1::port::PORT_COUNT_REAR;
     pub use crate::v1::port::PortCreateParams;
+    pub use crate::v1::port::PortFec;
+    pub use crate::v1::port::PortId;
     pub use crate::v1::port::PortIdPathParams;
     pub use crate::v1::port::PortIpv4Path;
     pub use crate::v1::port::PortIpv6Path;
+    pub use crate::v1::port::PortMedia;
     pub use crate::v1::port::PortSettings;
     pub use crate::v1::port::PortSettingsTag;
+    pub use crate::v1::port::PortSpeed;
     pub use crate::v1::port::PortToken;
+    pub use crate::v1::port::QsfpPort;
+    pub use crate::v1::port::RearPort;
+    pub use crate::v1::port::TxEq;
+    pub use crate::v1::port::TxEqSwHw;
+
+    pub use crate::v12::port::PortPrbsMode;
 }
 
 pub mod port_map {
@@ -173,19 +216,19 @@ pub mod snapshot {
 }
 
 pub mod switch_identifiers {
-    pub use crate::v1::switch_identifiers::ChipRevision;
-    pub use crate::v1::switch_identifiers::DisabledFeatures;
-    pub use crate::v1::switch_identifiers::FrequencySettings;
-    pub use crate::v1::switch_identifiers::FuseData;
-    pub use crate::v1::switch_identifiers::ManufacturingData;
-    pub use crate::v1::switch_identifiers::PartInfo;
-
+    pub use crate::v10::switch_identifiers::ChipRevision;
+    pub use crate::v10::switch_identifiers::DisabledFeatures;
+    pub use crate::v10::switch_identifiers::FrequencySettings;
+    pub use crate::v10::switch_identifiers::FuseData;
+    pub use crate::v10::switch_identifiers::ManufacturingData;
+    pub use crate::v10::switch_identifiers::PartInfo;
     pub use crate::v10::switch_identifiers::SwitchIdentifiers;
 }
 
 pub mod switch_port {
     pub use crate::v1::switch_port::Led;
     pub use crate::v1::switch_port::LedPolicy;
+    pub use crate::v1::switch_port::LedState;
     pub use crate::v1::switch_port::ManagementMode;
     pub use crate::v1::switch_port::SwitchPortView;
 }
@@ -199,12 +242,46 @@ pub mod table {
     pub use crate::v1::table::TableDumpResult;
     pub use crate::v1::table::TableEntry;
     pub use crate::v1::table::TableParam;
+
+    pub use crate::impls::table::TableEntryAction;
+    pub use crate::impls::table::TableEntryKey;
 }
 
 pub mod transceivers {
+    pub use crate::v1::transceivers::ActiveCableMediaInterfaceId;
+    pub use crate::v1::transceivers::ApplicationDescriptor;
+    pub use crate::v1::transceivers::Aux1Monitor;
+    pub use crate::v1::transceivers::Aux2Monitor;
+    pub use crate::v1::transceivers::Aux3Monitor;
+    pub use crate::v1::transceivers::AuxMonitors;
+    pub use crate::v1::transceivers::BaseTMediaInterfaceId;
+    pub use crate::v1::transceivers::CmisDatapath;
+    pub use crate::v1::transceivers::CmisDatapathState;
+    pub use crate::v1::transceivers::CmisLaneStatus;
+    pub use crate::v1::transceivers::ConnectorType;
+    pub use crate::v1::transceivers::Datapath;
     pub use crate::v1::transceivers::ElectricalMode;
+    pub use crate::v1::transceivers::EthernetComplianceCode;
+    pub use crate::v1::transceivers::ExtendedSpecificationComplianceCode;
     pub use crate::v1::transceivers::FaultReason;
+    pub use crate::v1::transceivers::HostElectricalInterfaceId;
+    pub use crate::v1::transceivers::Identifier;
+    pub use crate::v1::transceivers::LanePolarity;
+    pub use crate::v1::transceivers::MediaInterfaceId;
+    pub use crate::v1::transceivers::MmfMediaInterfaceId;
+    pub use crate::v1::transceivers::Monitors;
+    pub use crate::v1::transceivers::Oui;
+    pub use crate::v1::transceivers::OutputStatus;
+    pub use crate::v1::transceivers::PassiveCopperMediaInterfaceId;
+    pub use crate::v1::transceivers::PowerMode;
+    pub use crate::v1::transceivers::PowerState;
     pub use crate::v1::transceivers::QsfpDevice;
+    pub use crate::v1::transceivers::ReceiverPower;
+    pub use crate::v1::transceivers::Sff8636Datapath;
+    pub use crate::v1::transceivers::SffComplianceCode;
+    pub use crate::v1::transceivers::SmfMediaInterfaceId;
     pub use crate::v1::transceivers::Transceiver;
     pub use crate::v1::transceivers::TransceiverInfo;
+    pub use crate::v1::transceivers::Vendor;
+    pub use crate::v1::transceivers::VendorInfo;
 }
